Introduce the `scan` interface
We define the interface of the `scan` function and add
test to verify its behavior.

The `scan` function is currently implemented as a for loop to
keep each PR size small. In the next PR, we will replace the
forward pass with an optimization implementation that lowers to
an HLO `While` op.